don't overwrite existing package.json file

+++ create-schnack/index.js
@@ -8,14 +8,7 @@ const CWD = process.env.INIT_CWD || process.cwd();
 let tag = process.argv.find(arg => arg.includes('--tag')) || '--tag=latest';
 tag = tag.split('=')[1];
 
-const pkg = {
-    name: 'schnack.example.com',
-    version: '1.0.0',
-    private: true,
-    scripts: {
-        start: 'schnack'
-    }
-};
+
 
 async function main() {
     const configPath = path.join(CWD, 'schnack.json');
@@ -32,15 +25,26 @@ async function main() {
         process.exit(1);
     }
 
-    const { plugins } = require(configPath);
+    const { plugins, schnack_host } = require(configPath);
+
+    // create package.json file if it doesn't exist
+    if (!fs.existsSync(path.join(CWD, 'package.json'))) {
+        const pkg = {
+            name: schnack_host.replace(/https?:\/\//, ''),
+            version: '1.0.0',
+            private: true,
+            scripts: {
+                start: 'schnack'
+            }
+        };
+        fs.writeFileSync(path.join(CWD, 'package.json'), JSON.stringify(pkg, null, 4), {
+            encoding: 'utf-8'
+        });
+    }
 
     const packages = Object.keys(plugins || {})
         .filter(id => id !== 'notify-webpush')
         .map(id => plugins[id].pkg || `@schnack/plugin-${id}`);
-
-    fs.writeFileSync(path.join(CWD, 'package.json'), JSON.stringify(pkg, null, 4), {
-        encoding: 'utf-8'
-    });
 
     console.log('[npm] Start package installation.');
     const npm = spawn('npm', ['install', '-SE', '--production', `schnack@${tag}`].concat(packages));
